lumo-app/API/requirements.txt
boto3==1.21.6
botocore==1.24.6
click==8.0.4
DateTime==4.4
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
zope.interface==5.4.0
